+++ pgpcr/gpg_ops.py
@@ -3,24 +3,29 @@ import gpg
 class GPGKey:
 
 	def __init__(self, home):
-		self.ctx = gpg.Context(home_dir=home)
-		self.masteralgo = "rsa4096"
-		self.subalgo = "rsa2048"
+		self._ctx = gpg.Context(home_dir=home)
+		self._masteralgo = "rsa4096"
+		self._subalgo = "rsa2048"
 
 	def genmaster(self, userid):
-		genkey = self.ctx.create_key(userid, algorithm=self.masteralgo, sign=True, certify=True, passphrase=True)
-		self.master = self.ctx.get_key(genkey.fpr)
+		genkey = self._ctx.create_key(userid, algorithm=self._masteralgo, sign=True, certify=True, passphrase=True)
+		self._master = self._ctx.get_key(genkey.fpr)
 
 	def gensub(self):
-		self.subsig = self.ctx.create_subkey(self.master, algorithm=self.subalgo, sign=True)
-		self.subenc = self.ctx.create_subkey(self.master, algorithm=self.subalgo, encrypt=True)
-		self.subauth = self.ctx.create_subkey(self.master, algorithm=self.subalgo, authenticate=True)
+		self._subsig = self._ctx.create_subkey(self._master, algorithm=self._subalgo, sign=True)
+		self._subenc = self._ctx.create_subkey(self._master, algorithm=self._subalgo, encrypt=True)
+		self._subauth = self._ctx.create_subkey(self._master, algorithm=self._subalgo, authenticate=True)
 
-	def set_progress(self, progress, hook=None):
-		self.ctx.set_progress_cb(progress, hook)
+	def setprogress(self, progress, hook=None):
+		self._ctx.set_progress_cb(progress, hook)
 
-	def set_password(self, password, hook=None):
-		self.ctx.set_passphrase_cb(password, hook)
+	def setpassword(self, password, hook=None):
+		self._ctx.set_passphrase_cb(password, hook)
 
+	def setalgorithms(self, master, sub):
+		if master is not None:
+			self._masteralgo = master
+		if sub is not None:
+			self._subalgo = sub
 
-error = gpg.errors.GPGMEError
+GPGMEError = gpg.errors.GPGMEError